refactor(tests): split server_test.go into focused modules
Replaces the 1816-line monolithic server_test.go with 9 focused
files, each under the 500-line policy limit, following the same
pattern used for scheduler_test.go (#198). Removes the grandfathered
exception for internal/server/server_test.go from file_size.py.

New files:
- server_helpers_test.go       — shared test infrastructure
- server_workflow_submit_test.go — workflow run submission
- server_workflow_status_test.go — status, list, step metadata
- server_workflow_delete_test.go — delete, lifecycle, path traversal
- server_agent_test.go         — agent CRUD handlers
- server_middleware_test.go    — constructor, request ID, panic recovery, shutdown
- server_errors_test.go        — 500 error paths (failingStore/failingRegistry)
- server_concurrent_test.go    — concurrent and mixed stress tests
- server_stubs_test.go         — stub endpoint tests
