+++ FILEMAP.md
@@ -1,13 +1,13 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-25  
-> **413** tracked files
+> **421** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
-| `internal/` | 100 |
+| `internal/` | 108 |
 | `docs/` | 81 |
 | `agents/` | 77 |
 | `tests/` | 65 |
@@ -27,8 +27,8 @@
 | Extension | Count |
 |-----------|-------|
 | `.py` | 153 |
+| `.go` | 105 |
 | `.md` | 97 |
-| `.go` | 97 |
 | `.yaml` | 25 |
 | `.rs` | 9 |
 | `.yml` | 6 |
@@ -422,7 +422,15 @@
 │   │   ├── logs_stream_handler_test.go
 │   │   ├── middleware.go
 │   │   ├── server.go
-│   │   ├── server_test.go
+│   │   ├── server_agent_test.go
+│   │   ├── server_concurrent_test.go
+│   │   ├── server_errors_test.go
+│   │   ├── server_helpers_test.go
+│   │   ├── server_middleware_test.go
+│   │   ├── server_stubs_test.go
+│   │   ├── server_workflow_delete_test.go
+│   │   ├── server_workflow_status_test.go
+│   │   ├── server_workflow_submit_test.go
 │   │   ├── types.go
 │   │   └── workflow_handlers.go
 │   └── state/
